agetty: split out credential loading to credentials.c
Move agetty_load_credentials() and its helpers (cred_read_str,
cred_read_num, cred_read_bool) to a new credentials.c file.
