+++ agetty-cmd/agetty.h
@@ -61,4 +61,6 @@ extern void agetty_log_err(const char *, ...) __attribute__((__noreturn__))
 extern void agetty_log_warn(const char *, ...)
 				__attribute__((__format__(printf, 1, 2)));
 
+extern void agetty_load_credentials(struct agetty_options *op);
+
 #endif /* UTIL_LINUX_AGETTY_H */
